pyproject.toml
[tool.poetry]
name = "turing-schedule"
version = "0.1.0"
description = ""
authors = ["Qiyas <qiyasturk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
